Tratamento de erros dos controllers

+++ src/main/java/br/com/meuprojeto/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.meuprojeto.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -59,30 +60,43 @@ public class TopicosController {
 	}
 	
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
+	public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
+		Optional<Topico> optional = topicoRepository.findById(id);
 		
-		Topico topico = topicoRepository.getOne(id);
+		if(optional.isPresent()) {
+			return ResponseEntity.ok(new DetalhesDoTopicoDto(optional.get())); // get() pega o real tópico dentro do optional
+		}
 		
-		return new DetalhesDoTopicoDto(topico);
+		return ResponseEntity.notFound().build();
+		
 	}
 	
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
+		Optional<Topico> optional = topicoRepository.findById(id);
 		
-		Topico topico = form.atualizar(id, topicoRepository);
-		topicoRepository.save(topico);
+		if(optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+			topicoRepository.save(topico);
+	
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
 		
-		return ResponseEntity.ok(new TopicoDto(topico));
+		return ResponseEntity.notFound().build();
 	}
 	
 	@DeleteMapping("{/id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id){
-		
-		topicoRepository.deleteById(id);
-		
-		return ResponseEntity.ok().build();
+		Optional<Topico> optional = topicoRepository.findById(id);
 
+		if(optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			
+			return ResponseEntity.ok().build();
+		}
+		
+		return ResponseEntity.notFound().build();
 	}
 }
